requirements.txt
streamlit~=1.27.2
langchain~=0.0.350
python-dotenv~=1.0.0
